chore: prepare VIP-609 for proposal
